requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.0.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Minify==0.37
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lesscpy==0.15.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
ply==3.11
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==0.19.2
pytz==2024.1
rcssmin==1.1.2
referencing==0.35.1
replicate==0.31.0
requests==2.32.3
rpds-py==0.20.0
sniffio==1.3.1
SQLAlchemy==2.0.21
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.3.7
WTForms==3.0.1
xxhash==3.4.1
